Move MessageEventDefinitionParser back to the right package
