+++ activiti-core/activiti-bpmn-converter/src/main/java/org/activiti/bpmn/converter/util/BpmnXMLUtil.java
@@ -15,7 +15,7 @@ import javax.xml.stream.XMLStreamWriter;
 
 import org.activiti.bpmn.constants.BpmnXMLConstants;
 import org.activiti.bpmn.converter.child.*;
-import org.activiti.bpmn.converter.child.multi.instance.MessageEventDefinitionParser;
+import org.activiti.bpmn.converter.child.MessageEventDefinitionParser;
 import org.activiti.bpmn.converter.child.multi.instance.MultiInstanceParser;
 import org.activiti.bpmn.model.*;
 import org.apache.commons.lang3.StringUtils;
